requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.7.2
python-dotenv==0.10.3
PyYAML==5.2
requests==2.22.0
slackclient==2.5.0
soupsieve==1.9.5
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
yarl==1.4.2
